[FT-22] fix: use Integer for transaction id to support SQLite autoincrement
SQLite only auto-increments INTEGER PRIMARY KEY (rowid alias). BIGINT
PRIMARY KEY is treated as a regular NOT NULL column with no sequence,
causing an IntegrityError on every insert against the SQLite dev database.

+++ alembic/versions/2026_04_13_1734-d82b574a14fc_create_transaction_table.py
@@ -25,7 +25,7 @@ def upgrade() -> None:
 
     op.create_table(
         "financial_transaction",
-        sa.Column("id", sa.BigInteger, primary_key=True, autoincrement=True),
+        sa.Column("id", sa.Integer, primary_key=True, autoincrement=True),
         sa.Column("amount", sa.Numeric(10, 2), nullable=False),
         sa.Column(
             "kind",
